[gfx] Rename debug_utils.hpp/cpp to vk_debug_utils
It's quite a common name like device/instance/...,
avoid possible filename collisions.
