+++ include/CMakeLists.txt
@@ -82,12 +82,12 @@ set(VOXEN_HEADERS
 	include/voxen/common/threadpool.hpp
 	include/voxen/common/uid.hpp
 	include/voxen/common/world_state.hpp
-	include/voxen/gfx/vk/debug_utils.hpp
 	include/voxen/gfx/vk/render_graph.hpp
 	include/voxen/gfx/vk/render_graph_builder.hpp
 	include/voxen/gfx/vk/render_graph_execution.hpp
 	include/voxen/gfx/vk/render_graph_resource.hpp
 	include/voxen/gfx/vk/render_graph_runner.hpp
+	include/voxen/gfx/vk/vk_debug_utils.hpp
 	include/voxen/gfx/vk/vk_instance.hpp
 	include/voxen/server/world.hpp
 	include/voxen/util/aabb.hpp
