+++ src/CMakeLists.txt
@@ -75,13 +75,13 @@ set(VOXEN_SOURCES
 	src/util/hash.cpp
 	src/util/log.cpp
 	src/voxen/common/uid.cpp
-	src/voxen/gfx/vk/debug_utils.cpp
 	src/voxen/gfx/vk/render_graph_builder.cpp
 	src/voxen/gfx/vk/render_graph_execution.cpp
 	src/voxen/gfx/vk/render_graph_private.cpp
 	src/voxen/gfx/vk/render_graph_private.hpp
 	src/voxen/gfx/vk/render_graph_resource.cpp
 	src/voxen/gfx/vk/render_graph_runner.cpp
+	src/voxen/gfx/vk/vk_debug_utils.cpp
 	src/voxen/gfx/vk/vk_instance.cpp
 )
